neurips2020/requirement.txt
pytorch==1.4.0
gensim==3.8.0
numpy==1.18.1
scipy==1.3.2
pickle==1.5.0
